+++ src/three/TilesRenderer.js
@@ -1,16 +1,12 @@
 import { TilesRendererBase } from '../base/TilesRendererBase.js';
-import { WGS84_HEIGHT, WGS84_RADIUS } from '../base/constants.js';
 import { B3DMLoader } from './B3DMLoader.js';
 import { PNTSLoader } from './PNTSLoader.js';
 import { I3DMLoader } from './I3DMLoader.js';
 import { CMPTLoader } from './CMPTLoader.js';
 import { GLTFExtensionLoader } from './GLTFExtensionLoader.js';
 import { TilesGroup } from './TilesGroup.js';
-import { EllipsoidRegion } from './math/EllipsoidRegion.js';
 import {
 	Matrix4,
-	Box3,
-	Sphere,
 	Vector3,
 	Vector2,
 	Frustum,
@@ -18,14 +14,12 @@ import {
 } from 'three';
 import { raycastTraverse, raycastTraverseFirstHit } from './raycastTraverse.js';
 import { readMagicBytes } from '../utilities/readMagicBytes.js';
+import { TileBoundingVolume } from './math/TileBoundingVolume.js';
 
 const INITIAL_FRUSTUM_CULLED = Symbol( 'INITIAL_FRUSTUM_CULLED' );
 const tempMat = new Matrix4();
 const tempMat2 = new Matrix4();
 const tempVector = new Vector3();
-const vecX = new Vector3();
-const vecY = new Vector3();
-const vecZ = new Vector3();
 
 const X_AXIS = new Vector3( 1, 0, 0 );
 const Y_AXIS = new Vector3( 0, 1, 0 );
@@ -112,7 +106,7 @@ export class TilesRenderer extends TilesRendererBase {
 	}
 
 	/* Public API */
-	getBounds( box ) {
+	getBounds( target ) {
 
 		if ( ! this.root ) {
 
@@ -120,26 +114,21 @@ export class TilesRenderer extends TilesRendererBase {
 
 		}
 
-		const cached = this.root.cached;
-		const boundingBox = cached.box;
-		const obbMat = cached.boxTransform;
+		const boundingVolume = this.root.cached.boundingVolume;
+		if ( boundingVolume ) {
 
-		if ( boundingBox ) {
-
-			box.copy( boundingBox );
-			box.applyMatrix4( obbMat );
-
+			boundingVolume.getAABB( target );
 			return true;
 
 		} else {
 
-			return false;
+			return true;
 
 		}
 
 	}
 
-	getOrientedBounds( box, matrix ) {
+	getOrientedBounds( targetBox, targetMatrix ) {
 
 		if ( ! this.root ) {
 
@@ -147,26 +136,21 @@ export class TilesRenderer extends TilesRendererBase {
 
 		}
 
-		const cached = this.root.cached;
-		const boundingBox = cached.box;
-		const obbMat = cached.boxTransform;
+		const boundingVolume = this.root.cached.boundingVolume;
+		if ( boundingVolume ) {
 
-		if ( boundingBox ) {
-
-			box.copy( boundingBox );
-			matrix.copy( obbMat );
-
+			boundingVolume.getOBB( targetBox, targetMatrix );
 			return true;
 
 		} else {
 
-			return false;
+			return true;
 
 		}
 
 	}
 
-	getBoundingSphere( sphere ) {
+	getBoundingSphere( target ) {
 
 		if ( ! this.root ) {
 
@@ -174,11 +158,10 @@ export class TilesRenderer extends TilesRendererBase {
 
 		}
 
-		const boundingSphere = this.root.cached.sphere;
+		const boundingVolume = this.root.cached.boundingVolume;
+		if ( boundingVolume ) {
 
-		if ( boundingSphere ) {
-
-			sphere.copy( boundingSphere );
+			boundingVolume.getSphere( target );
 			return true;
 
 		} else {
@@ -214,7 +197,7 @@ export class TilesRenderer extends TilesRendererBase {
 
 		if ( raycaster.firstHitOnly ) {
 
-			const hit = raycastTraverseFirstHit( this.root, this.group, this.activeTiles, raycaster );
+			const hit = raycastTraverseFirstHit( this, this.root, raycaster );
 			if ( hit ) {
 
 				intersects.push( hit );
@@ -223,7 +206,7 @@ export class TilesRenderer extends TilesRendererBase {
 
 		} else {
 
-			raycastTraverse( this.root, this.group, this.activeTiles, raycaster, intersects );
+			raycastTraverse( this, this.root, raycaster, intersects );
 
 		}
 
@@ -433,9 +416,9 @@ export class TilesRenderer extends TilesRendererBase {
 
 	}
 
-	preprocessNode( tile, parentTile, tileSetDir ) {
+	preprocessNode( tile, tileSetDir, parentTile = null ) {
 
-		super.preprocessNode( tile, parentTile, tileSetDir );
+		super.preprocessNode( tile, tileSetDir, parentTile );
 
 		const transform = new Matrix4();
 		if ( tile.transform ) {
@@ -460,117 +443,24 @@ export class TilesRenderer extends TilesRendererBase {
 		}
 
 		const transformInverse = new Matrix4().copy( transform ).invert();
-
-		let box = null;
-		let boxTransform = null;
-		let boxTransformInverse = null;
-		if ( 'box' in tile.boundingVolume ) {
-
-			const data = tile.boundingVolume.box;
-			box = new Box3();
-			boxTransform = new Matrix4();
-			boxTransformInverse = new Matrix4();
-
-			// get the extents of the bounds in each axis
-			vecX.set( data[ 3 ], data[ 4 ], data[ 5 ] );
-			vecY.set( data[ 6 ], data[ 7 ], data[ 8 ] );
-			vecZ.set( data[ 9 ], data[ 10 ], data[ 11 ] );
-
-			const scaleX = vecX.length();
-			const scaleY = vecY.length();
-			const scaleZ = vecZ.length();
-
-			vecX.normalize();
-			vecY.normalize();
-			vecZ.normalize();
-
-			// handle the case where the box has a dimension of 0 in one axis
-			if ( scaleX === 0 ) {
-
-				vecX.crossVectors( vecY, vecZ );
-
-			}
-
-			if ( scaleY === 0 ) {
-
-				vecY.crossVectors( vecX, vecZ );
-
-			}
-
-			if ( scaleZ === 0 ) {
-
-				vecZ.crossVectors( vecX, vecY );
-
-			}
-
-			// create the oriented frame that the box exists in
-			boxTransform.set(
-				vecX.x, vecY.x, vecZ.x, data[ 0 ],
-				vecX.y, vecY.y, vecZ.y, data[ 1 ],
-				vecX.z, vecY.z, vecZ.z, data[ 2 ],
-				0, 0, 0, 1
-			);
-			boxTransform.premultiply( transform );
-			boxTransformInverse.copy( boxTransform ).invert();
-
-			// scale the box by the extents
-			box.min.set( - scaleX, - scaleY, - scaleZ );
-			box.max.set( scaleX, scaleY, scaleZ );
-
-		}
-
-		let sphere = null;
+		const boundingVolume = new TileBoundingVolume();
 		if ( 'sphere' in tile.boundingVolume ) {
 
-			const data = tile.boundingVolume.sphere;
-			sphere = new Sphere();
-			sphere.center.set( data[ 0 ], data[ 1 ], data[ 2 ] );
-			sphere.radius = data[ 3 ];
-			sphere.applyMatrix4( transform );
-
-		} else if ( 'box' in tile.boundingVolume ) {
-
-			const data = tile.boundingVolume.box;
-			sphere = new Sphere();
-			box.getBoundingSphere( sphere );
-			sphere.center.set( data[ 0 ], data[ 1 ], data[ 2 ] );
-			sphere.applyMatrix4( transform );
+			boundingVolume.setSphereData( ...tile.boundingVolume.sphere, transform );
 
 		}
 
-		let region = null;
+		if ( 'box' in tile.boundingVolume ) {
+
+			boundingVolume.setObbData( tile.boundingVolume.box, transform );
+
+		}
+
 		if ( 'region' in tile.boundingVolume ) {
 
-			const data = tile.boundingVolume.region;
-			const [ west, south, east, north, minHeight, maxHeight ] = data;
-
-			region = new EllipsoidRegion(
-				WGS84_RADIUS, WGS84_RADIUS, WGS84_HEIGHT,
-				south, north,
-				west, east,
-				minHeight, maxHeight,
-			);
-
-			if ( sphere === null ) {
-
-				sphere = new Sphere();
-				region.getBoundingSphere( sphere );
-
-			}
-
-			if ( box === null ) {
-
-				box = new Box3();
-				boxTransform = new Matrix4();
-				boxTransformInverse = new Matrix4();
-
-				region.getBoundingBox( box, boxTransform );
-				boxTransformInverse.copy( boxTransform ).invert();
-
-			}
+			boundingVolume.setRegionData( ...tile.boundingVolume.region );
 
 		}
-
 
 		tile.cached = {
 
@@ -581,11 +471,7 @@ export class TilesRenderer extends TilesRendererBase {
 			active: false,
 			inFrustum: [],
 
-			box,
-			boxTransform,
-			boxTransformInverse,
-			sphere,
-			region,
+			boundingVolume,
 
 			scene: null,
 			geometry: null,
@@ -614,23 +500,24 @@ export class TilesRenderer extends TilesRendererBase {
 		const cached = tile.cached;
 		const cachedTransform = cached.transform;
 
+		const upAdjustment = new Matrix4();
 		switch ( upAxis.toLowerCase() ) {
 
 			case 'x':
-				tempMat.makeRotationAxis( Y_AXIS, - Math.PI / 2 );
+				upAdjustment.makeRotationAxis( Y_AXIS, - Math.PI / 2 );
 				break;
 
 			case 'y':
-				tempMat.makeRotationAxis( X_AXIS, Math.PI / 2 );
+				upAdjustment.makeRotationAxis( X_AXIS, Math.PI / 2 );
 				break;
 
 			case 'z':
-				tempMat.identity();
+				upAdjustment.identity();
 				break;
 
 		}
 
-		const fileType = readMagicBytes( buffer ) || extension;
+		const fileType = ( readMagicBytes( buffer ) || extension ).toLowerCase();
 		switch ( fileType ) {
 
 			case 'b3dm': {
@@ -639,11 +526,9 @@ export class TilesRenderer extends TilesRendererBase {
 				loader.workingPath = workingPath;
 				loader.fetchOptions = fetchOptions;
 
-				loader.adjustmentTransform.copy( tempMat );
+				loader.adjustmentTransform.copy( upAdjustment );
 
-				promise = loader
-					.parse( buffer )
-					.then( res => res.scene );
+				promise = loader.parse( buffer );
 				break;
 
 			}
@@ -653,9 +538,7 @@ export class TilesRenderer extends TilesRendererBase {
 				const loader = new PNTSLoader( manager );
 				loader.workingPath = workingPath;
 				loader.fetchOptions = fetchOptions;
-				promise = loader
-					.parse( buffer )
-					.then( res => res.scene );
+				promise = loader.parse( buffer );
 				break;
 
 			}
@@ -666,11 +549,9 @@ export class TilesRenderer extends TilesRendererBase {
 				loader.workingPath = workingPath;
 				loader.fetchOptions = fetchOptions;
 
-				loader.adjustmentTransform.copy( tempMat );
+				loader.adjustmentTransform.copy( upAdjustment );
 
-				promise = loader
-					.parse( buffer )
-					.then( res => res.scene );
+				promise = loader.parse( buffer );
 				break;
 
 			}
@@ -681,7 +562,7 @@ export class TilesRenderer extends TilesRendererBase {
 				loader.workingPath = workingPath;
 				loader.fetchOptions = fetchOptions;
 
-				loader.adjustmentTransform.copy( tempMat );
+				loader.adjustmentTransform.copy( upAdjustment );
 
 				promise = loader
 					.parse( buffer )
@@ -696,9 +577,7 @@ export class TilesRenderer extends TilesRendererBase {
 				const loader = new GLTFExtensionLoader( manager );
 				loader.workingPath = workingPath;
 				loader.fetchOptions = fetchOptions;
-				promise = loader
-					.parse( buffer )
-					.then( res => res.scene	);
+				promise = loader.parse( buffer );
 				break;
 
 			default:
@@ -708,7 +587,21 @@ export class TilesRenderer extends TilesRendererBase {
 
 		}
 
-		return promise.then( scene => {
+		return promise.then( result => {
+
+			let scene;
+			let metadata;
+			if ( result.isObject3D ) {
+
+				scene = result;
+				metadata = null;
+
+			} else {
+
+				scene = result.scene;
+				metadata = result;
+
+			}
 
 			if ( tile._loadIndex !== loadIndex ) {
 
@@ -725,7 +618,7 @@ export class TilesRenderer extends TilesRendererBase {
 			// rotation fix which is why "multiply" happens here.
 			if ( fileType === 'glb' || fileType === 'gltf' ) {
 
-				scene.matrix.multiply( tempMat );
+				scene.matrix.multiply( upAdjustment );
 
 			}
 
@@ -737,8 +630,6 @@ export class TilesRenderer extends TilesRendererBase {
 
 			} );
 			updateFrustumCulled( scene, ! this.autoDisableRendererCulling );
-
-			cached.scene = scene;
 
 			// We handle raycasting in a custom way so remove it from here
 			scene.traverse( c => {
@@ -781,6 +672,8 @@ export class TilesRenderer extends TilesRendererBase {
 			cached.materials = materials;
 			cached.geometry = geometry;
 			cached.textures = textures;
+			cached.scene = scene;
+			cached.metadata = metadata;
 
 			if ( this.onLoadModel ) {
 
@@ -838,6 +731,7 @@ export class TilesRenderer extends TilesRendererBase {
 			cached.materials = null;
 			cached.textures = null;
 			cached.geometry = null;
+			cached.metadata = null;
 
 		}
 
@@ -894,15 +788,7 @@ export class TilesRenderer extends TilesRendererBase {
 		const inFrustum = cached.inFrustum;
 		const cameras = this.cameras;
 		const cameraInfo = this.cameraInfo;
-
-		// TODO: Use the content bounding volume here?
-		// TODO: We should use the largest distance to the tile between
-		// all available bounding volume types.
-		const boundingSphere = cached.sphere;
-		const boundingBox = cached.box;
-		const boxTransformInverse = cached.boxTransformInverse;
-		const transformInverse = cached.transformInverse;
-		const useBox = boundingBox && boxTransformInverse;
+		const boundingVolume = cached.boundingVolume;
 
 		let maxError = - Infinity;
 		let minDistance = Infinity;
@@ -927,24 +813,7 @@ export class TilesRenderer extends TilesRendererBase {
 
 			} else {
 
-				tempVector.copy( info.position );
-
-				let distance;
-				if ( useBox ) {
-
-					tempVector.applyMatrix4( boxTransformInverse );
-					distance = boundingBox.distanceToPoint( tempVector );
-
-				} else {
-
-					tempVector.applyMatrix4( transformInverse );
-					// Sphere#distanceToPoint is negative inside the sphere, whereas Box3#distanceToPoint is
-					// zero inside the box. Clipping the distance to a minimum of zero ensures that both
-					// types of bounding volume behave the same way.
-					distance = Math.max( boundingSphere.distanceToPoint( tempVector ), 0 );
-
-				}
-
+				const distance = boundingVolume.distanceToPoint( info.position );
 				const scaledDistance = distance * invScale;
 				const sseDenominator = info.sseDenominator;
 				error = tile.geometricError / ( scaledDistance * sseDenominator );
@@ -964,40 +833,30 @@ export class TilesRenderer extends TilesRendererBase {
 
 	tileInView( tile ) {
 
-		// TODO: we should use the more precise bounding volumes here if possible
-		// cache the root-space planes
-		// Use separating axis theorem for frustum and obb
-
 		const cached = tile.cached;
-		const sphere = cached.sphere;
+		const boundingVolume = cached.boundingVolume;
 		const inFrustum = cached.inFrustum;
-		if ( sphere ) {
+		const cameraInfo = this.cameraInfo;
+		let inView = false;
+		for ( let i = 0, l = cameraInfo.length; i < l; i ++ ) {
 
-			const cameraInfo = this.cameraInfo;
-			let inView = false;
-			for ( let i = 0, l = cameraInfo.length; i < l; i ++ ) {
+			// Track which camera frustums this tile is in so we can use it
+			// to ignore the error calculations for cameras that can't see it
+			const frustum = cameraInfo[ i ].frustum;
+			if ( boundingVolume.intersectsFrustum( frustum ) ) {
 
-				// Track which camera frustums this tile is in so we can use it
-				// to ignore the error calculations for cameras that can't see it
-				const frustum = cameraInfo[ i ].frustum;
-				if ( frustum.intersectsSphere( sphere ) ) {
+				inView = true;
+				inFrustum[ i ] = true;
 
-					inView = true;
-					inFrustum[ i ] = true;
+			} else {
 
-				} else {
-
-					inFrustum[ i ] = false;
-
-				}
+				inFrustum[ i ] = false;
 
 			}
 
-			return inView;
-
 		}
 
-		return true;
+		return inView;
 
 	}